Add "minimal" "working" Gradle/JNI/AndroidStudio example
This commit contains an Android app project, based on
https://codelabs.developers.google.com/codelabs/android-studio-jni/index.html
that uses

* Android Studio as its IDE,
* Gradle as its build toolkit, and
* JNI to interface the Java/app part with some C code.

It's not great. It builds however (given you manage to install the
required toolchain and adjust your `local.properties` properly),
and shows that the Java part can launch the C part which in turn
can call back into a Java method to poll a function, and lastly deliver
that function's return value back to the calling piece of Java code.

(Those skilled in the art will recognize conceptual similarity with
@SensibilityTestbed and aaaaalbert/py_jni_sensors. Indeed, we are
prototyping stuff here that should be used in these other projects
later on.)